Refactor out some certificate acess functions to separate file
